Refactored logic from Mentors.js
