+++ src/screens/Mentors.js
@@ -24,122 +24,20 @@ import { H1, H2, H3, H4, H6, P } from '../components/Text';
 import Toast from 'react-native-simple-toast';
 import moment from 'moment';
 
+import * as ScreenLogic from 'utils/ScreenLogic'
+
 const serverURL = "https://technicamentorshipservertest.herokuapp.com"
 
 export default class Mentors extends Component<Props> {
   constructor(props) {
     super(props);
     this.state = { appState: AppState.currentState, question: '', location: "", newQuestionScreen:false, listData: [] };
-    this.sendQuestion = this.sendQuestion.bind(this);
-    this.showToast = this.showToast.bind(this);
-    this._handleAppStateChange = this._handleAppStateChange.bind(this);
+    this.sendQuestion = ScreenLogic.sendQuestion.bind(this);
+    this.showToast = ScreenLogic.showToast.bind(this);
+    this._handleAppStateChange = ScreenLogic._handleAppStateChange.bind(this);
   }
 
-  grabQuestionsFromDB(email) {
-    fetch(`${serverURL}/getquestions/${email}`, {
-      method: 'GET',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
-      },
-    }).then(response => response.json()).then(async (responseJson) => {
-      console.log("questions found")
-      console.log(responseJson)
-      this.setState({listData: responseJson });
-  }).catch(err => {
-    console.log("ERROR GRABBING QUESTIONS")
-    console.log(err)
-  })
 
-  }
-
-  // initially loads question data from server
-  async componentDidMount() {
-    AppState.addEventListener('change', this._handleAppStateChange)
-    const user_data = await AsyncStorage.getItem("USER_DATA_STORE");
-    const user_data_json = JSON.parse(user_data)
-    this.grabQuestionsFromDB(user_data_json.user_data.email)
-  }
-
-  componentWillUnmount() {
-    AppState.removeEventListener('change', this._handleAppStateChange);
-  }
-
-  async _handleAppStateChange(nextAppState){
-    if (this.state.appState.match(/inactive|background/) && nextAppState === 'active') {
-      console.log('App has come to the foreground!')
-      const user_data = await AsyncStorage.getItem("USER_DATA_STORE");
-      const user_data_json = JSON.parse(user_data)
-      this.grabQuestionsFromDB(user_data_json.user_data.email)
-    }
-    this.setState({appState: nextAppState});
-  }
-
-  clearInputs() {
-    this.setState({ question: '', location: ''});
-  }
-  cancelQuestion() {
-    this.setState({ question: '', newQuestionScreen: !this.state.newQuestionScreen });
-  }
-  toggleModal() {
-    this.setState({ newQuestionScreen: !this.state.newQuestionScreen });
-  }
-
-  showToast() {
-    // Show toast after 600ms
-    // This 600ms delay ensures the toast loads after the modal animation close
-    // happens. There is a weird iOS issue where toast will vanish the moment
-    // modal closes. This is the best workaround I could make for now.
-    setTimeout(() => {
-      Toast.show('Question sent! Our next available mentor will come assist you.', Toast.LONG);
-    }, 400)
-  }
-
-  async sendQuestion() {
-    if (this.state.question === '' || this.state.location === '') {
-      Alert.alert(
-        "Try Again",
-        "Your question or location was empty.",
-        [
-          {text: 'OK', onPress: () => console.log('OK Pressed')},
-        ],
-        { cancelable: false }
-      );
-    } else {
-      const fcmToken = await AsyncStorage.getItem("FCMToken");
-      const user_data = await AsyncStorage.getItem("USER_DATA_STORE");
-      const user_data_json = JSON.parse(user_data);
-      const name = user_data_json.user_data.first_name + " " + user_data_json.user_data.last_name
-      var questionObject = {
-        question: this.state.question,
-        location: this.state.location,
-        status: "Awaiting available mentors",
-        key: moment().format(),
-        name: name,
-        email: user_data_json.user_data.email,
-      }
-      if (fcmToken != null) {
-        questionObject.fcmToken = fcmToken
-      }
-
-      var questionString = JSON.stringify(questionObject)
-      fetch(`${serverURL}/question`, {
-        method: 'POST',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: questionString,
-      }).catch(error => {
-        console.log(error)
-      })
-      this.clearInputs()
-      this.showToast();
-      this.toggleModal()
-      // make new question show up immediately at top of list
-      this.setState({listData: [questionObject].concat(this.state.listData)})
-    }
-  }
   renderHeading() {
     return (
       <React.Fragment>
@@ -210,14 +108,14 @@ export default class Mentors extends Component<Props> {
           </View>
         </View>
         <View marginTop = {10}>
-        <TouchableOpacity onPress={() => this.sendQuestion()}>
+        <TouchableOpacity onPress={() => ScreenLogic.sendQuestion()}>
           <Button
             text="Submit Question"
           />
         </TouchableOpacity>
         </View>
         <View marginTop = {10}>
-          <TouchableOpacity onPress={() => this.cancelQuestion()}>
+          <TouchableOpacity onPress={() => ScreenLogic.cancelQuestion()}>
             <Button
               text="Cancel"
             />
@@ -226,43 +124,6 @@ export default class Mentors extends Component<Props> {
       </Modal>
     )
   }
-
-  async createNotificationListener() {
-    // updates when app is in foreground
-    this.notificationListener = firebase
-    .notifications()
-    .onNotification(notification => {this.grabQuestionsFromDB(notification.data.email)});
-
-    // updates when app is in the background
-    this.notificationOpenedListener = firebase.notifications().onNotificationOpened((notificationOpen) => {
-      this.grabQuestionsFromDB(notificationOpen.notification.data.email)
-    });
-  }
-
-  async updateQuestionStatus(notification) {
-    console.log('notification received', notification.body);
-    console.log(notification.data)
-
-    const key = notification.data.key;
-    const mentorName = notification.data.mentor_name;
-
-    const questions = await AsyncStorage.getItem("questions");
-    const qList = JSON.parse(questions);
-
-    // update status of question
-    qList.forEach((element, index) => {
-      if (element.key == key) {
-        console.log("found!");
-        element.status = `${mentorName} has claimed your question!`;
-        qList[index] = element;
-      }
-    })
-    // store update in local storage
-    await AsyncStorage.setItem("questions", JSON.stringify(qList))
-    this.setState({listData: qList})
-  }
-
-
 
   render() {
     { this.createNotificationListener() }
